refactor(cli): move output formatting into parser_cli_core::output
